biocXqtl: modular system for creating molecular QTL statistics

This is a modular system for creating molecular QTL statistics. Given a RangedSummarizedExperiment with assay elements comprised of molecular responses like RNA-seq counts, and genotype calls in colData, results of association tests will be bound into the rowData of the input RangedSummarizedExperiment.

See also

Author

Maintainer: Vince Carey stvjc@channing.harvard.edu (ORCID)